requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0.8
django-autotranslate==1.1.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-modeltranslation==0.15.1
django-rosetta==0.9.4
django-taggit==1.3.0
dynaconf==3.0.0
futures==3.1.1
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
goslate==1.5.1
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
packaging==20.9
pathspec==0.8.0
Pillow==7.2.0
polib==1.1.0
protobuf==3.15.6
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-box==3.4.6
python-dotenv==0.12.0
python-gettext==4.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
rsa==4.7.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.1.0
